"Extend PGParams functionality and add documentation"
The commit further expands the functionality of the PGParams class, specifically adding more methods like addLike() and addEqualNullable() for dealing with 'LIKE' and nullable values in SQL queries respectively. In addition, the functionality of existing methods like add() and replaceSQLWithValues() is also enhanced. This update will greatly improve the flexibility and ease of use in dealing with PostgreSQL queries.

Moreover, detailed in-line documentation is added for all the functions in PGParams. This will significantly aid in understanding the purpose and usage of the class and its methods.